app/requirements.txt
fastapi[all]
uvicorn
pydantic
scalar-fastapi
sqlmodel
pydantic-settings
asyncpg
passlib[bcrypt]
pyjwt
redis
alembic
